refactor form questions out of giant constants.ts
