+++ laprogram/app/feedback/components/sections/EndOfQuarterSection.tsx
@@ -1,11 +1,7 @@
 "use client";
 
 import type { FeedbackFormValues } from "../../schema";
-import {
-  END_OF_QUARTER_NONSENSITIVE_QUESTIONS,
-  EQ_SENSITIVE_QUESTIONS,
-  EQ_NONSENSITIVE_TEXT_FIELDS,
-} from "../../constants";
+
 import { LikertField } from "../fields/LikertField";
 import { TextareaFormField } from "../fields/TextareaFormField";
 import { FieldGroup } from "@/components/ui/field";
@@ -13,6 +9,11 @@ import { withForm } from "../../form";
 import { defaultValues, feedbackFormSchema } from "../../schema";
 import { ActivitiesField } from "../fields/ActivitiesField";
 import { HoursField } from "../fields/HoursField";
+import {
+  EQ_NONSENSITIVE_QUESTIONS,
+  EQ_SENSITIVE_QUESTIONS,
+  EQ_NONSENSITIVE_TEXT_FIELDS,
+} from "../../questions/end_quarter";
 
 export const EndOfQuarterSection = withForm({
   defaultValues,
@@ -21,17 +22,15 @@ export const EndOfQuarterSection = withForm({
     <FieldGroup>
       <ActivitiesField form={form} />
       <HoursField form={form} />
-      {END_OF_QUARTER_NONSENSITIVE_QUESTIONS.map(
-        ({ value, label, options }) => (
-          <LikertField
-            key={value}
-            form={form}
-            fieldName={value as keyof FeedbackFormValues}
-            label={label}
-            options={options!}
-          />
-        ),
-      )}
+      {EQ_NONSENSITIVE_QUESTIONS.map(({ value, label, options }) => (
+        <LikertField
+          key={value}
+          form={form}
+          fieldName={value as keyof FeedbackFormValues}
+          label={label}
+          options={options!}
+        />
+      ))}
       {EQ_SENSITIVE_QUESTIONS.map(({ value, label, options }) => (
         <LikertField
           key={value}
