+++ laprogram/app/feedback/components/sections/MidQuarterSection.tsx
@@ -1,11 +1,6 @@
 "use client";
 
 import type { FeedbackFormValues } from "../../schema";
-import {
-  MID_QUARTER_NONSENSITIVE_QUESTIONS,
-  MQ_NONSENSITIVE_TEXT_FIELDS,
-  MQ_SENSITIVE_TEXT_FIELDS,
-} from "../../constants";
 import { ActivitiesField } from "../fields/ActivitiesField";
 import { HoursField } from "../fields/HoursField";
 import { LikertField } from "../fields/LikertField";
@@ -13,6 +8,11 @@ import { TextareaFormField } from "../fields/TextareaFormField";
 import { FieldGroup } from "@/components/ui/field";
 import { withForm } from "../../form";
 import { defaultValues, feedbackFormSchema } from "../../schema";
+import {
+  MQ_NONSENSITIVE_QUESTIONS,
+  MQ_NONSENSITIVE_TEXT_FIELDS,
+  MQ_SENSITIVE_TEXT_FIELDS,
+} from "../../questions/mid_quarter";
 
 export const MidQuarterSection = withForm({
   defaultValues,
@@ -21,7 +21,7 @@ export const MidQuarterSection = withForm({
     <FieldGroup>
       <ActivitiesField form={form} />
       <HoursField form={form} />
-      {MID_QUARTER_NONSENSITIVE_QUESTIONS.map(({ value, label, options }) => (
+      {MQ_NONSENSITIVE_QUESTIONS.map(({ value, label, options }) => (
         <LikertField
           key={value}
           form={form}
