+++ laprogram/app/feedback/components/sections/TASection.tsx
@@ -1,15 +1,15 @@
 "use client";
 
 import type { FeedbackFormValues } from "../../schema";
-import {
-  TA_NONSENSITIVE_QUESTIONS,
-  TA_NONSENSITIVE_TEXT_FIELDS,
-} from "../../constants";
 import { LikertField } from "../fields/LikertField";
 import { TextareaFormField } from "../fields/TextareaFormField";
 import { FieldGroup } from "@/components/ui/field";
 import { withForm } from "../../form";
 import { defaultValues, feedbackFormSchema } from "../../schema";
+import {
+  TA_NONSENSITIVE_QUESTIONS,
+  TA_NONSENSITIVE_TEXT_FIELDS,
+} from "../../questions/ta";
 
 export const TASection = withForm({
   defaultValues,
